Reorganised repo around functional groupings rather than module type.
