+++ package.json
@@ -3,32 +3,32 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "start": "MODE=development CONFIG=dev webpack serve --open",
+    "start": "cross-env MODE=development CONFIG=dev webpack serve --open",
     "build": "webpack",
-    "build:dev": "MODE=development CONFIG=dev npm run build",
-    "build:test-staging": "MODE=production CONFIG=test-staging npm run build",
-    "build:beta": "MODE=production CONFIG=beta npm run build",
-    "build:zm-staging": "MODE=production CONFIG=zm-staging npm run build",
-    "build:zm-prod": "MODE=production CONFIG=zm-prod npm run build",
-    "build:za-staging": "MODE=production CONFIG=za-staging npm run build",
-    "build:za-prod": "MODE=production CONFIG=za-prod npm run build",
-    "build:et-staging": "MODE=production CONFIG=et-staging npm run build",
-    "build:et-prod": "MODE=production CONFIG=et-prod npm run build",
-    "build:mw-staging": "MODE=production CONFIG=mw-staging npm run build",
-    "build:mw-prod": "MODE=production CONFIG=mw-prod npm run build",
-    "build:jm-staging": "MODE=production CONFIG=jm-staging npm run build",
-    "build:jm-prod": "MODE=production CONFIG=jm-prod npm run build",
-    "build:ca-staging": "MODE=production CONFIG=ca-staging npm run build",
-    "build:uk-staging": "MODE=production CONFIG=uk-staging npm run build",
-    "build:ro-staging": "MODE=production CONFIG=ro-staging npm run build",
-    "build:hu-staging": "MODE=production CONFIG=hu-staging npm run build",
-    "build:hu-prod": "MODE=production CONFIG=hu-prod npm run build",
-    "build:e2e-dev": "MODE=development CONFIG=e2e-dev npm run build",
-    "build:co-staging": "MODE=production CONFIG=co-staging npm run build",
-    "build:co-production": "MODE=production CONFIG=co-production npm run build",
-    "build:cl-staging": "MODE=production CONFIG=cl-staging npm run build",
-    "build:zw-staging": "MODE=production CONFIG=zw-staging npm run build",
-    "build:pl-staging": "MODE=production CONFIG=pl-staging npm run build",
+    "build:dev": "cross-env MODE=development CONFIG=dev npm run build",
+    "build:test-staging": "cross-env MODE=production CONFIG=test-staging npm run build",
+    "build:beta": "cross-env MODE=production CONFIG=beta npm run build",
+    "build:zm-staging": "cross-env MODE=production CONFIG=zm-staging npm run build",
+    "build:zm-prod": "cross-env MODE=production CONFIG=zm-prod npm run build",
+    "build:za-staging": "cross-env MODE=production CONFIG=za-staging npm run build",
+    "build:za-prod": "cross-env MODE=production CONFIG=za-prod npm run build",
+    "build:et-staging": "cross-env MODE=production CONFIG=et-staging npm run build",
+    "build:et-prod": "cross-env MODE=production CONFIG=et-prod npm run build",
+    "build:mw-staging": "cross-env MODE=production CONFIG=mw-staging npm run build",
+    "build:mw-prod": "cross-env MODE=production CONFIG=mw-prod npm run build",
+    "build:jm-staging": "cross-env MODE=production CONFIG=jm-staging npm run build",
+    "build:jm-prod": "cross-env MODE=production CONFIG=jm-prod npm run build",
+    "build:ca-staging": "cross-env MODE=production CONFIG=ca-staging npm run build",
+    "build:uk-staging": "cross-env MODE=production CONFIG=uk-staging npm run build",
+    "build:ro-staging": "cross-env MODE=production CONFIG=ro-staging npm run build",
+    "build:hu-staging": "cross-env MODE=production CONFIG=hu-staging npm run build",
+    "build:hu-prod": "cross-env MODE=production CONFIG=hu-prod npm run build",
+    "build:e2e-dev": "cross-env MODE=development CONFIG=e2e-dev npm run build",
+    "build:co-staging": "cross-env MODE=production CONFIG=co-staging npm run build",
+    "build:co-production": "cross-env MODE=production CONFIG=co-production npm run build",
+    "build:cl-staging": "cross-env MODE=production CONFIG=cl-staging npm run build",
+    "build:zw-staging": "cross-env MODE=production CONFIG=zw-staging npm run build",
+    "build:pl-staging": "cross-env MODE=production CONFIG=pl-staging npm run build",
     "lint": "eslint --ext js --ext jsx --ext ts --ext tsx src/",
     "lint:fix": "npm run lint -- --fix"
   },
@@ -40,6 +40,7 @@
     "@typescript-eslint/parser": "^5.37.0",
     "babel-loader": "^8.2.5",
     "babel-runtime": "^6.26.0",
+    "cross-env": "^7.0.3",
     "eslint": "^8.23.1",
     "eslint-config-twilio-react": "^2.0.0",
     "html-webpack-plugin": "^5.0.0",
