+++ src/index.ts
@@ -1,3 +1,5 @@
 import { initWebchat } from './aselo-webchat';
 
 initWebchat();
+export {ExitButtonBase} from "./end-chat/end-chat-styles";
+export {EndButtonBase} from "./end-chat/end-chat-styles";
